% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnPowerlaw.R
\name{AnPowerlaw}
\alias{AnPowerlaw}
\title{return power spectral density vector for a power law with variance 1}
\usage{
AnPowerlaw(beta, freq)
}
\arguments{
\item{beta}{}

\item{freq}{frequency vector}
}
\value{
power spectral density
}
\author{
Thomas Laepple
}

