% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{calcE}
\alias{calcE}
\title{Calculate the Eccentricity anomaly}
\usage{
calcE(M, e, tol = 1e-08)
}
\arguments{
\item{M}{mean anomaly}

\item{tol}{tolerance}
}
\value{
Eccentricity anomaly E
}
\description{
based on
Grossman; The sheer joy of celestial mechanics
Orbits under the inverse square law
}
\details{
Calculate the Eccentricity anomaly E from the mean anomaly M, iterativly
inversion of Keplers formula: E-e*sin(E)=M  (188)
}
\author{
Thomas Laepple
}
