% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{calcf}
\alias{calcf}
\title{Calculate the true anomaly}
\usage{
calcf(E, e)
}
\arguments{
\item{E}{}

\item{e}{}
}
\value{
true anomaly f (theta)
}
\description{
Calculate the true anomaly f (theta) from the eccentricity anomaly
true anomaly = angle between the point and perihelion
based on
Grossman; The sheer joy of celestial mechanics
Orbits under the inverse square law
book: (181)
}
\author{
Thomas Laepple
}

