% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{DailyInsolation}
\alias{DailyInsolation}
\title{Computes the daily average insolation for any time in the last 5Million years}
\usage{
DailyInsolation(kyear, lat, day, day_type = 1, fast = TRUE, T.alpha = 80,
  alpha = 0)
}
\arguments{
\item{kyear}{Thousands of years before present (0 to 5000).}

\item{lat}{Latitude in degrees (-90 to 90).}

\item{day}{Indicator of time of year; calendar day by default.}

\item{day_type}{Convention for specifying time of year (1,2,3) [optional], see details}

\item{fast}{TRUE = use precalculated values of orbital parameters in 100yr steps}

\item{T.alpha}{for custom calendar (day_type=3)}

\item{alpha}{for custom calendar (day_type=3)}
}
\value{
list(Fsw=Fsw,ecc=ecc,obliquity=obliquity,long_perh=long_perh,lambda=lambda/2/pi*360 )
  Fsw = Daily average solar radiation in W/m^2.
  ... and all orbital parameters
}
\description{
Computes the daily average insolation, converted and adapted from the Eisenmann and Huybers,2006
Matlab Code
Orbital parameters from Berger and Loutre 1991
calculation of daily average insolation following Berger A. (1978).
Iterative solution for the Kepler equation, following Grossman; "The sheer joy of celestial mechanics,
Orbits under the inverse square law"
}
\details{
Authors of original MATLAB version :
  Ian Eisenman and Peter Huybers, Harvard University, August 2006
  eisenman@fas.harvard.edu
  This file is available online at
  http://deas.harvard.edu/~eisenman/downloads

Translation into R and extension to arbitray calendars; Thomas Laepple, 2008
Details
 day_type: Convention for specifying time of year (1,2,3) [optional].
    day_type=1 (default): day input is calendar day (1-365.24), where day 1
      is January first.  The calendar is referenced to the vernal equinox
      which always occurs at day 80.
    day_type=2: day input is solar longitude (0-360 degrees). Solar
      longitude is the angle of the Earth's orbit measured from spring
      equinox (21 March). Note that calendar days and solar longitude are
      not linearly related because, by Kepler's Second Law, Earth's
      angular velocity varies according to its distance from the sun.
 day_type=3; Custom calendar
   day input is calendar day (1-365.24), where day 1
   is January first.  The calendar is referenced to the solar longitude alpha
   (0..2pi, angle of the earth orbit from spring equinox)
      which always occurs at day T.alpha.
this uses an iterative solution of Keplers equation and is therefore
much slower than the day_type=1 which uses a trigeometric approximation
of Kepler's equation
}
\examples{
par(mfcol=c(2,1))
plot(0:1000,DailyInsolation(0:1000,40,180)$Fsw,main="day 180 (= peak summer) insolation at 40degN",xlab="kyr BP",ylab="W/m2",type="l")
plot(1:365,DailyInsolation(0,40,1:365)$Fsw,main="modern daily insolation at 40degN",xlab="day of the year",ylab="W/m2",type="l")
}
\author{
Thomas Laepple
}
