% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM-Derivative-Functions.R
\name{Deriv}
\alias{Deriv}
\title{Deriv}
\source{
https://gist.github.com/gavinsimpson/e73f011fdaaab4bb5a30
}
\usage{
Deriv(mod, n = 200, eps = 1e-07, newdata, term)
}
\arguments{
\item{mod}{A fitted GAM or GAMM}

\item{n}{The number of points at which to evaluate the derivative, defaults to 200}

\item{eps}{The finite difference in x (delta) over which to estimate change in y}

\item{newdata}{A dataframe of points (+ levels of any factors in the model) at which to evaluate the derivatives}

\item{term}{The term in the GAM(M) over which which derivatives should be estimated}
}
\description{
Deriv
}
\author{
Gavin Simpson
}
