% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{f.from.t}
\alias{f.from.t}
\title{true angle from days}
\usage{
f.from.t(ecc, t)
}
\arguments{
\item{ecc}{eccentricity}

\item{t}{days from perihelion}
}
\value{
true angle from perihelion (via the mean angle M)
}
\description{
based on
Grossman; The sheer joy of celestial mechanics
Orbits under the inverse square law
book: (181)
}
\author{
Thomas Laepple
}

