% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs-pField-methods.R
\name{gethistory}
\alias{gethistory}
\title{gethistory}
\usage{
gethistory(data)
}
\arguments{
\item{data}{a pTs or pField object}
}
\value{
character
}
\description{
returns the "history" attribute of a pTs/pField object
}
\seealso{
Other history: \code{\link{addhistory}}
}

