% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs-pField-methods.R
\name{GetName}
\alias{GetName}
\alias{getname}
\title{GetName}
\usage{
GetName(data)
}
\arguments{
\item{data}{a pTs or pField object}
}
\value{
character
}
\description{
returns the "name" attribute of a pTs/pField object
}
\examples{
getname(c(a = 2))
GetName(c(a = 2))
}
\author{
Thomas Laepple
}
