% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetCDFAtCoords.R
\name{GetNetCDFAtCoords}
\alias{GetNetCDFAtCoords}
\title{Extract timeseries for nearest coordinates from a netcdf file}
\usage{
GetNetCDFAtCoords(filename, req.coords, req.var, time.var = "time")
}
\arguments{
\item{req.coords}{matrix or dataframe of longitude and latitude coordinates}

\item{req.var}{name of the variable to extract}

\item{time.var}{name of the time variable}
}
\description{
Extract timeseries for nearest coordinates from a netcdf file
}
