% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggortho.R
\name{ggortho}
\alias{ggortho}
\title{Plot a ggplot map in orthographic projection}
\usage{
ggortho(
  centre.lat,
  centre.lon,
  lat.ax.vals = NULL,
  long.ax.vals = NULL,
  lat.ax.labs.pos = NULL,
  ocean.fill = "aliceblue",
  land.fill = "grey92",
  border.colour = "grey55",
  lon.lat.colour = "grey55",
  lon.lat.linetype = 4,
  bury.lon.lat.lines = FALSE,
  plot.visible = FALSE
)
}
\arguments{
\item{centre.lat}{Centre map on this latitude}

\item{centre.lon}{Centre map on this longitude}

\item{lat.ax.vals}{Optionally specify values for latitude labels}

\item{long.ax.vals}{Optionally specify values for longitude labels}

\item{lat.ax.labs.pos}{Specify the line of longitude on which the latitude
labels will be printed}

\item{ocean.fill}{Specify the colour of the ocean}

\item{land.fill}{Specify the colour of the land}

\item{border.colour}{Specify the colour of the land borders}

\item{lon.lat.colour}{Colour of the longitude and latitude lines}

\item{lon.lat.linetype}{Linetype of the longitude and latitude lines}

\item{bury.lon.lat.lines}{Hide longitude and latitude lines under land}

\item{plot.visible}{Return a rectangular projection showing the area that 
will be visible in the orthographic plot}
}
\value{
ggplot object
}
\description{
Plot a ggplot map in orthographic projection
}
\details{
Adapted from code here:
https://gist.github.com/fzenoni/ef23faf6d1ada5e4a91c9ef23b0ba2c1
}
\examples{
# Centre on 90 degrees East, 45 degrees north
p <- ggortho(centre.lat = 45, centre.lon = 90,
            lat.ax.vals = seq(-90, 90, 30))
p

# Hide lon lat lines on land
p <- ggortho(centre.lat = 45, centre.lon = 90,
            lat.ax.vals = seq(-90, 90, 30),
            bury.lon.lat.lines = TRUE)
p



# Plot Antarctica and add a point
p <- ggortho(centre.lat = -90, centre.lon = 0,
            lat.ax.vals = seq(-90, 90, 30),
            lat.ax.labs.pos = 60)

test.dat <-
 data.frame(lat = c(-80), lon = c(100))

test.dat.sf <- sf::st_as_sf(test.dat, coords = c("lon", "lat")) \%>\%
 sf::st_set_crs(4326)

p + ggplot2::geom_sf(data = test.dat.sf, colour = "Red")


# change the colour of the land

p <- ggortho(centre.lat = -30, centre.lon = -120,
            lat.ax.vals = seq(-90, 90, 30),
            #lat.ax.labs.pos = 60,
            plot.visible = FALSE,
            land.fill = "Darkgreen")

test.dat <-
 data.frame(lat = c(-35:-36), lon = c(-120:-121))

test.dat.sf <- sf::st_as_sf(test.dat, coords = c("lon", "lat")) \%>\%
 sf::st_set_crs(4326)

p + ggplot2::geom_sf(data = test.dat.sf, colour = "Red")
}
