% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolar.R
\name{ggpolar}
\alias{ggpolar}
\title{Nicely formatted polar ggplot}
\usage{
ggpolar(pole = c("N", "S"), max.lat, min.lat, longitude.spacing = 60,
  land.fill.colour = "Grey", n.lat.labels = 4, nearest.x.degrees = 5)
}
\arguments{
\item{pole}{which pole N or S?}

\item{max.lat}{maximum latitude}

\item{min.lat}{minimum latitude}

\item{longitude.spacing}{interval between longitude axis tick marks}

\item{land.fill.colour}{colour to shade the land}

\item{n.lat.labels}{approximate number of latitude tickmarks}

\item{nearest.x.degrees}{round latitude tickmarks to how many degrees?}
}
\description{
Produce a nicely formatted map of the Arctic or Antarctic in
polar projection. Adapted from Mikey Harper's plot
https://stackoverflow.com/a/49084793
}
\examples{
ggpolar(pole = "N", max.lat = 90, min.lat = 55, n.lat.labels = 4)
ggpolar(pole = "S", max.lat = -60, min.lat = -90)
\dontrun{
ggpolar(pole = "W", max.lat = -55, min.lat = -90)
ggpolar(pole = "S", max.lat = 90, min.lat = 55)
ggpolar(pole = "S", max.lat = -90, min.lat = -55)
}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
