% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pField.R
\name{pField}
\alias{pField}
\title{pField}
\source{
Function copied from "proxytype.R" in paleolibary/src/
}
\usage{
pField(data, time, lat = 0, lon = 0, name = " ", history = " ",
  date = TRUE)
}
\arguments{
\item{data}{input data, usually a 3d array}

\item{time}{vector}

\item{lat}{vector of latitudes, defaults to 0}

\item{lon}{vector of longitudes, defaults to 0}

\item{name}{character string giving the name of the pField object to be generated}

\item{history}{character string to append to the history attribute}

\item{date}{logical, whether or not to append the current date to the history attribute}
}
\value{
a pField object
}
\description{
adds attributes such as longitudes and latitudes to a gridded
  data array (which in most cases is three-dimensional and where the dimensions
  stand in general for longitude, latitude and time) and assigns the class
  "pField" to the resulting object.
}
\author{
Thomas Laepple
}
