% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM-Derivative-Functions.R
\name{plot.Deriv}
\alias{plot.Deriv}
\title{Title}
\usage{
\method{plot}{Deriv}(x, alpha = 0.05, polygon = TRUE, sizer = FALSE,
  term, eval = 0, lwd = 3, col = "lightgrey", border = col, ylab,
  xlab, main, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
