% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-functions.R
\name{SDofSD}
\alias{SDofSD}
\title{Standard deviation (error) of a standard deviation}
\usage{
SDofSD(s, n)
}
\arguments{
\item{s}{standard deviation}

\item{n}{number of samples from which s was estimated}
}
\value{
numeric
}
\description{
Calculate the standard deviation (i.e. standard error) of a
standard deviation, given n, the number of samples from which the standard
deviation was estimated.
}
\details{
if n <= 0 an exact method is used, otherwise Stirling's
 approximation is used to avoid numerical errors calculating the gamma
 function for large values of n. Formulae taken from
 https://stats.stackexchange.com/q/28567
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
