% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{t.from.f}
\alias{t.from.f}
\title{days to true angle}
\usage{
\method{t}{from.f}(ecc, f)
}
\arguments{
\item{ecc}{eccentricity}

\item{f}{true angle from perihelion}
}
\value{
days from perihelion, calculated using the mean angle M
}
\description{
based on
Grossman; The sheer joy of celestial mechanics
Orbits under the inverse square law
book: (181)
}
\author{
Thomas Laepple
}

