% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{unwrap}
\alias{unwrap}
\title{corrects the radian phase angles in array P by adding multiples of <U+00B1>2pi when absolute jumps between consecutive array elements are greater than pi radians.}
\usage{
unwrap(p)
}
\arguments{
\item{p}{radian phase angles array}
}
\value{
unwrapped angles
}
\description{
Q = unwrap(P) corrects the radian phase angles in array P by adding multiples of <U+00B1>2pi when absolute jumps between consecutive array elements are greater than pi radians.
based on http://ccrma.stanford.edu/~jos/sasp/Matlab_listing_unwrap_m.html
}
\author{
Thomas Laepple
}

