% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs-pField-methods.R
\name{AddHistory}
\alias{AddHistory}
\alias{addhistory}
\title{AddHistory}
\usage{
AddHistory(x, newhist)
}
\arguments{
\item{x}{a pTs or pField object}

\item{newhist}{character string}
}
\value{
a pTs or pField object
}
\description{
adds a history (e.g., a comment) to a pTs or a pField object
}
\seealso{
Other history: \code{\link{GetHistory}}
}
\author{
Thomas Laepple
}
