% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs-pField-methods.R
\name{GetHistory}
\alias{GetHistory}
\alias{gethistory}
\title{GetHistory}
\usage{
GetHistory(data)
}
\arguments{
\item{data}{a pTs or pField object}
}
\value{
character
}
\description{
returns the "history" attribute of a pTs/pField object
}
\seealso{
Other history: \code{\link{AddHistory}}
}
\author{
Thomas Laepple
}
