% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs-pField-methods.R
\name{getname}
\alias{getname}
\title{getname}
\usage{
getname(data)
}
\arguments{
\item{data}{a pTs or pField object}
}
\value{
character
}
\description{
returns the "name" attribute of a pTs/pField object
}

